+++ models/thompson2020/tests/model_test.py
@@ -21,7 +21,7 @@ class TestThompson202Models(unittest.TestCase):
         assert_dicts_approx_equal(expected_macro, actual_macro)
         assert len(expected_micro) == len(actual_micro)
         for expected, actual in zip(expected_micro, actual_micro):
-            assert_dicts_approx_equal(expected, actual)
+            assert_dicts_approx_equal(expected, actual, abs=1e-4)
 
     @parameterized.expand(get_testing_device_parameters())
     def test_prism_sources_example(self, device: int):
@@ -38,7 +38,7 @@ class TestThompson202Models(unittest.TestCase):
         assert_dicts_approx_equal(expected_macro, actual_macro)
         assert len(expected_micro) == len(actual_micro)
         for expected, actual in zip(expected_micro, actual_micro):
-            assert_dicts_approx_equal(expected, actual)
+            assert_dicts_approx_equal(expected, actual, abs=1e-4)
 
     def test_invalid_input(self):
         model = Prism()
@@ -70,3 +70,21 @@ class TestThompson202Models(unittest.TestCase):
         ]
         with self.assertRaises(Exception):
             model.predict_batch(inputs)
+
+    @parameterized.expand(get_testing_device_parameters())
+    def test_prism_translation_regression(self, device: int):
+        # Tests using Prism as a translation system using
+        # the examples from https://github.com/thompsonb/prism/tree/master/translation
+        model = Prism(device=device)
+        inputs = [
+            {"source": "Hi world."},
+            {"source": "This is a Test."},
+            {"source": "Some of my Best Friends are Linguists."},
+        ]
+
+        actual_outputs = model.translate_batch("fr", inputs)
+        assert actual_outputs == [
+            "Le Monde.",
+            "C'est un test.",
+            "Certains de mes meilleurs amis sont linguistes.",
+        ]