+++ Changelog.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added [BLANC](models/vasilyev2020/Readme.md)
 - Added [METEOR](models/denkowski2014/Readme.md)
 - Added a role question generator from [Pyatkin et al. (2021)](models/pyatkin2021/Readme.md)
+- Added using [Prism](models/thompson2020/Readme.md) as an MT model
 
 ### Fixed
 - Fixed an error in [Lite3Pyramid](models/zhang2021/Readme.md) by updating to a newer version of the code.